+++ packages/app/src/App.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import { Navigate, Route } from 'react-router-dom';
+import { Route } from 'react-router-dom';
 import { apiDocsPlugin, ApiExplorerPage } from '@backstage/plugin-api-docs';
+import { HomepageCompositionRoot } from '@backstage/plugin-home';
+import { homePage } from './components/home/HomePage';
 import {
   CatalogEntityPage,
   CatalogIndexPage,
@@ -79,7 +81,9 @@ const app = createApp({
 
 const routes = (
   <FlatRoutes>
-    <Route path="/" element={<Navigate to="catalog" />} />
+    <Route path="/" element={<HomepageCompositionRoot />}>
+      {homePage}
+    </Route>
     <Route path="/catalog" element={<CatalogIndexPage />} />
     <Route
       path="/catalog/:namespace/:kind/:name"