+++ plugins/jira-dashboard/src/hooks/useJira.ts
@@ -2,6 +2,12 @@ import useAsync from 'react-use/lib/useAsync';
 import { JiraResponse } from '@axis-backstage/plugin-jira-dashboard-common';
 import { JiraDashboardApi } from '../api';
 
+/**
+ * Hook to get the Jira data for a given entity.
+ * @param entityRef - The entity reference to get Jira data for
+ * @param jiraDashboardApi - The Jira dashboard API
+ * @public
+ */
 export function useJira(
   entityRef: string,
   jiraDashboardApi: JiraDashboardApi,